+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/expression/function/scalar/datetime/DateTrunc.java
@@ -5,6 +5,7 @@
  */
 package org.elasticsearch.xpack.sql.expression.function.scalar.datetime;
 
+import org.elasticsearch.common.time.IsoLocale;
 import org.elasticsearch.xpack.sql.expression.Expression;
 import org.elasticsearch.xpack.sql.expression.Expressions;
 import org.elasticsearch.xpack.sql.expression.Nullability;
@@ -14,8 +15,13 @@ import org.elasticsearch.xpack.sql.expression.gen.script.ScriptTemplate;
 import org.elasticsearch.xpack.sql.tree.NodeInfo;
 import org.elasticsearch.xpack.sql.tree.Source;
 import org.elasticsearch.xpack.sql.type.DataType;
+import org.elasticsearch.xpack.sql.util.StringUtils;
 
 import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.temporal.ChronoField;
+import java.util.HashSet;
+import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.Set;
@@ -45,6 +51,16 @@ public class DateTrunc extends BinaryScalarFunction {
         MICROSECOND("microseconds", "mcs"),
         NANOSECOND("nanoseconds", "ns");
 
+        private static Set<String> ALL_DATE_PARTS;
+
+        static {
+            ALL_DATE_PARTS = new HashSet<>();
+            for (DatePart datePart : DatePart.values()) {
+                ALL_DATE_PARTS.add(datePart.name().toLowerCase(IsoLocale.ROOT));
+                ALL_DATE_PARTS.addAll(datePart.aliases());
+            }
+        }
+
         private Set<String> aliases;
 
         DatePart(String... aliases) {
@@ -63,6 +79,101 @@ public class DateTrunc extends BinaryScalarFunction {
                 }
             }
             return null;
+        }
+
+        public static List<String> findSimilar(String match) {
+            return StringUtils.findSimilar(match, ALL_DATE_PARTS);
+        }
+
+        public static ZonedDateTime truncate(ZonedDateTime dateTime, DateTrunc.DatePart datePart) {
+            ZonedDateTime truncated = null;
+            switch (datePart) {
+                case MILLENNIUM:
+                    int year = dateTime.getYear();
+                    int firstYearOfMillenium = year - (year % 1000);
+                    truncated = dateTime
+                        .with(ChronoField.YEAR, firstYearOfMillenium)
+                        .with(ChronoField.MONTH_OF_YEAR, 1)
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case CENTURY:
+                    year = dateTime.getYear();
+                    int firstYearOfCentury = year - (year % 100);
+                    truncated = dateTime
+                        .with(ChronoField.YEAR, firstYearOfCentury)
+                        .with(ChronoField.MONTH_OF_YEAR, 1)
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case DECADE:
+                    year = dateTime.getYear();
+                    int firstYearOfDecade = year - (year % 10);
+                    truncated = dateTime
+                        .with(ChronoField.YEAR, firstYearOfDecade)
+                        .with(ChronoField.MONTH_OF_YEAR, 1)
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case YEAR:
+                    truncated = dateTime
+                        .with(ChronoField.MONTH_OF_YEAR, 1)
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case QUARTER:
+                    int month = dateTime.getMonthValue();
+                    int firstMonthOfQuarter = (((month - 1) / 3) * 3) + 1;
+                    truncated = dateTime
+                        .with(ChronoField.MONTH_OF_YEAR, firstMonthOfQuarter)
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case MONTH:
+                    truncated = dateTime
+                        .with(ChronoField.DAY_OF_MONTH, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case WEEK:
+                    truncated = dateTime
+                        .with(ChronoField.DAY_OF_WEEK, 1)
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case DAY:
+                    truncated = dateTime
+                        .toLocalDate().atStartOfDay(dateTime.getZone());
+                    break;
+                case HOUR:
+                    int hour = dateTime.getHour();
+                    truncated = dateTime.toLocalDate().atStartOfDay(dateTime.getZone())
+                        .with(ChronoField.HOUR_OF_DAY, hour);
+                    break;
+                case MINUTE:
+                    hour = dateTime.getHour();
+                    int minute = dateTime.getMinute();
+                    truncated = dateTime.toLocalDate().atStartOfDay(dateTime.getZone())
+                        .with(ChronoField.HOUR_OF_DAY, hour)
+                        .with(ChronoField.MINUTE_OF_HOUR, minute);
+                    break;
+                case SECOND:
+                    truncated = dateTime
+                        .with(ChronoField.NANO_OF_SECOND, 0);
+                    break;
+                case MILLISECOND:
+                    int micros = dateTime.get(ChronoField.MICRO_OF_SECOND);
+                    truncated = dateTime
+                        .with(ChronoField.MILLI_OF_SECOND, (micros / 1000));
+                    break;
+                case MICROSECOND:
+                    int nanos = dateTime.getNano();
+                    truncated = dateTime
+                        .with(ChronoField.MICRO_OF_SECOND, (nanos / 1000));
+                    break;
+                case NANOSECOND:
+                    truncated = dateTime;
+                    break;
+            }
+            return truncated;
         }
     }
 
@@ -86,11 +197,22 @@ public class DateTrunc extends BinaryScalarFunction {
             return resolution;
         }
 
-        if (left().foldable() && DatePart.resolveTruncate((String) left().fold()) == null) {
-            return new TypeResolution(format(null, "first argument of [{}] must be one of {} or their aliases, found value [{}]",
-                sourceText(),
-                DatePart.values(),
-                Expressions.name(left())));
+        if (left().foldable()) {
+            String truncateToValue = (String) left().fold();
+            if (DatePart.resolveTruncate(truncateToValue) == null) {
+                List<String> similar = DatePart.findSimilar(truncateToValue);
+                if (similar.isEmpty()) {
+                    return new TypeResolution(format(null, "first argument of [{}] must be one of {} or their aliases, found value [{}]",
+                        sourceText(),
+                        DatePart.values(),
+                        Expressions.name(left())));
+                } else {
+                    return new TypeResolution(format(null, "Unknown value [{}] for first argument of [{}]; did you mean {}?",
+                        Expressions.name(left()),
+                        sourceText(),
+                        similar));
+                }
+            }
         }
         resolution = isDate(right(), sourceText(), Expressions.ParamOrdinal.SECOND);
         if (resolution.unresolved()) {
@@ -137,6 +259,11 @@ public class DateTrunc extends BinaryScalarFunction {
     }
 
     @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), zoneId);
+    }
+
+    @Override
     public boolean equals(Object o) {
         if (this == o) {
             return true;
@@ -149,10 +276,5 @@ public class DateTrunc extends BinaryScalarFunction {
         }
         DateTrunc dateTrunc = (DateTrunc) o;
         return Objects.equals(zoneId, dateTrunc.zoneId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), zoneId);
     }
 }
